adds longview detail routing
