+++ packages/manager/src/features/Longview/LongviewDetail/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LongviewDetail';
