+++ packages/manager/src/features/Longview/LongviewLanding/LongviewLanding.tsx
@@ -1,51 +1,132 @@
 import * as React from 'react';
 import { RouteComponentProps } from 'react-router-dom';
 import { compose } from 'recompose';
-import setDocs from 'src/components/DocsSidebar/setDocs';
-import { DocumentTitleSegment } from 'src/components/DocumentTitle';
-import Placeholder from 'src/components/Placeholder';
-import { MonitoringYourServer } from 'src/documentation';
+import { makeStyles, Theme } from 'src/components/core/styles';
 
-import Landing from './LandingContent';
+import AddNewLink from 'src/components/AddNewLink';
+import Breadcrumb from 'src/components/Breadcrumb';
+import Box from 'src/components/core/Box';
+import Divider from 'src/components/core/Divider';
+import DocumentationButton from 'src/components/DocumentationButton';
+import Grid from 'src/components/Grid';
 
-import useFlags from 'src/hooks/useFlags';
+import withLongviewClients, {
+  Props as LongviewProps
+} from 'src/containers/longview.container';
 
-const docs: Linode.Doc[] = [
-  {
-    title: 'What is Longview and How to Use it',
-    src: 'https://www.linode.com/docs/platform/longview/longview/',
-    body: `This guide shows how to install and use Linode Longview.`
-  },
-  MonitoringYourServer
-];
+import AddClientDrawer from './AddClientDrawer';
+import DeleteDialog from './LongviewDeleteDialog';
+import LongviewTable from './LongviewTable';
+import UpdateDrawer from './UpdateClientDrawer';
 
-export const LongviewLanding: React.FC<RouteComponentProps> = props => {
-  const flags = useFlags();
-
-  if (flags.longview) {
-    return <Landing {...props} />;
+const useStyles = makeStyles((theme: Theme) => ({
+  line: {
+    marginTop: theme.spacing(10),
+    marginBottom: theme.spacing(3)
   }
+}));
+
+type CombinedProps = RouteComponentProps & LongviewProps;
+
+const LongviewContent: React.FC<CombinedProps> = props => {
+  const classes = useStyles();
+
+  const [addDrawerOpen, toggleAddDrawer] = React.useState<boolean>(false);
+  const [editDrawerOpen, toggleEditDrawer] = React.useState<boolean>(false);
+  const [deleteDialogOpen, toggleDeleteDialog] = React.useState<boolean>(false);
+  const [selectedClientID, setClientID] = React.useState<number | undefined>(
+    undefined
+  );
+  const [selectedClientLabel, setClientLabel] = React.useState<string>('');
+
+  React.useEffect(() => {
+    props.getLongviewClients();
+  }, []);
+
+  const openDeleteDialog = (id: number, label: string) => {
+    toggleDeleteDialog(true);
+    setClientID(id);
+    setClientLabel(label);
+  };
+
+  const openEditDrawer = (id: number, label: string) => {
+    toggleEditDrawer(true);
+    setClientID(id);
+    setClientLabel(label);
+  };
+
+  const {
+    longviewClientsData,
+    longviewClientsError,
+    longviewClientsLastUpdated,
+    longviewClientsLoading,
+    longviewClientsResults,
+    createLongviewClient,
+    deleteLongviewClient,
+    updateLongviewClient
+  } = props;
 
   return (
     <React.Fragment>
-      <DocumentTitleSegment segment="Longview" />
-      <Placeholder
-        title="Longview"
-        copy="Keep your Linux systems running smoothly with insights from your system metrics."
-        buttonProps={[
-          {
-            onClick: () =>
-              window.open(
-                'https://manager.linode.com/longview',
-                '_blank',
-                'noopener'
-              ),
-            children: 'Navigate to Classic Manager'
-          }
-        ]}
+      <Box display="flex" flexDirection="row" justifyContent="space-between">
+        <Breadcrumb pathname={props.location.pathname} labelTitle="Longview" />
+        <DocumentationButton href={'https://google.com'} />
+      </Box>
+      <Divider className={classes.line} type="landingHeader" />
+      <Grid
+        container
+        justify="flex-end"
+        alignItems="flex-end"
+        style={{ paddingBottom: 0 }}
+      >
+        <Grid item>
+          <Grid container alignItems="flex-end">
+            <Grid item className="pt0">
+              <AddNewLink
+                onClick={() => toggleAddDrawer(true)}
+                label="Add a Client"
+              />
+            </Grid>
+          </Grid>
+        </Grid>
+      </Grid>
+      <LongviewTable
+        longviewClientsData={longviewClientsData}
+        longviewClientsError={longviewClientsError}
+        longviewClientsLastUpdated={longviewClientsLastUpdated}
+        longviewClientsLoading={longviewClientsLoading}
+        longviewClientsResults={longviewClientsResults}
+        triggerDeleteLongviewClient={openDeleteDialog}
+        triggerEditLongviewClient={openEditDrawer}
+      />
+      <AddClientDrawer
+        title="Add Longview Client"
+        onClose={() => toggleAddDrawer(false)}
+        open={addDrawerOpen}
+        createClient={createLongviewClient}
+      />
+      <UpdateDrawer
+        title={`Rename Longview Client${
+          selectedClientLabel ? `: ${selectedClientLabel}` : ''
+        }`}
+        selectedID={selectedClientID}
+        selectedLabel={selectedClientLabel}
+        updateClient={updateLongviewClient}
+        open={editDrawerOpen}
+        onClose={() => toggleEditDrawer(false)}
+      />
+      <DeleteDialog
+        selectedLongviewClientID={selectedClientID}
+        selectedLongviewClientLabel={selectedClientLabel}
+        deleteClient={deleteLongviewClient}
+        open={deleteDialogOpen}
+        closeDialog={() => toggleDeleteDialog(false)}
       />
     </React.Fragment>
   );
 };
 
-export default compose<RouteComponentProps, {}>(setDocs(docs))(LongviewLanding);
+export default compose<CombinedProps, RouteComponentProps>(
+  React.memo,
+  withLongviewClients()
+)(LongviewContent);
