+++ packages/manager/src/features/Longview/index.tsx
@@ -1,34 +1,59 @@
 import * as React from 'react';
-import {
-  Route,
-  RouteComponentProps,
-  Switch,
-  withRouter
-} from 'react-router-dom';
+import { Route, RouteComponentProps, Switch } from 'react-router-dom';
 import DefaultLoader from 'src/components/DefaultLoader';
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
+import Placeholder from 'src/components/Placeholder';
+import useFlags from 'src/hooks/useFlags';
 
 const LongviewLanding = DefaultLoader({
   loader: () => import('./LongviewLanding')
 });
 
+const LongviewDetail = DefaultLoader({
+  loader: () => import('./LongviewDetail')
+});
+
 type Props = RouteComponentProps<{}>;
 
-class Longview extends React.Component<Props> {
-  render() {
-    const {
-      match: { path }
-    } = this.props;
+const Longview: React.FC<Props> = props => {
+  const {
+    match: { path }
+  } = props;
 
+  const flags = useFlags();
+
+  if (!flags.longview) {
     return (
       <React.Fragment>
         <DocumentTitleSegment segment="Longview" />
-        <Switch>
-          <Route component={LongviewLanding} path={path} exact />
-        </Switch>
+        <Placeholder
+          title="Longview"
+          copy="Keep your Linux systems running smoothly with insights from your system metrics."
+          buttonProps={[
+            {
+              onClick: () =>
+                window.open(
+                  'https://manager.linode.com/longview',
+                  '_blank',
+                  'noopener'
+                ),
+              children: 'Navigate to Classic Manager'
+            }
+          ]}
+        />
       </React.Fragment>
     );
   }
-}
 
-export default withRouter(Longview);
+  return (
+    <React.Fragment>
+      <DocumentTitleSegment segment="Longview" />
+      <Switch>
+        <Route component={LongviewDetail} path={`${path}/clients/:id`} />
+        <Route component={LongviewLanding} />
+      </Switch>
+    </React.Fragment>
+  );
+};
+
+export default Longview;
